requirements.txt
dj-database-url==0.5.0
Django==2.2.10
django-heroku==0.3.1
djangorestframework==3.10.3
ez-setup==0.9
gunicorn==20.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
